ai/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
python-dotenv==1.0.0
google-generativeai==0.8.3
google-cloud-texttospeech==2.18.0
pydantic==2.10.5
redis==5.2.1
psycopg2-binary==2.9.10
httpx==0.28.1
